requirements.txt
numpy>=1.23.2
pandas>=2.0.0
yfinance==0.1.63
pytz==2021.1
Flask>=2.2.5
flask-cors==4.0.0
gunicorn==20.1.0
Werkzeug>=2.2.5
supabase==1.0.3
python-dotenv==1.0.0
requests==2.31.0
APScheduler==3.10.4
SQLAlchemy>=2.0.28
PyMySQL==1.1.0
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.0.33
